requirement.txt
Django==1.5.1
South
PIL
sorl-thumbnail
django-tinymce==1.5.1
-e git://github.com/zvadim/django-imagefield.git#egg=advanced_imagefield
pytils
django-haystack
Whoosh
-e git+git@github.com:ImaginaryLandscape/django-filebrowser-no-grappelli.git#egg=filebrowser-no-grappelli
django-modeltranslation==0.6
django-rosetta==0.7.2